+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Repair kits now have different "efficiency" levels, giving more incentive to upgrade. Values are configurable.
+### Changed
+- 'server' config is now 'common' and will appear in the config folder, like most configs do
+
 ## [1.10.2] - 2020-07-23
 ### Added
 - (1.16.1) Coating part type